Refactors level initialization and grid building
Moves grid building and level initialization logic to use new events and data structures.

Replaces the old `LevelInitEvent` and payload with a `BuildGridEvent` that takes the level grid as input and a `BuildGridDoneEvent` that signals completion of grid construction.

Updates `GameGrid` to listen to `BuildGridEvent` and `LevelLoader` to raise it with level data.

Removes LevelInitPayload and updates level initialization in GamePlay to use the new LevelDatabase to populate GoalCount.

Updates CameraFollow to use LevelDatabase to access the grid origin, center and size.