Adds water nut and water hole interactions
Adds new `WaterNut` and `WaterHoleGoal` types to the game.

Implements the interaction where a `WaterNut` falling into a `HoleGoal` transforms the hole into a `WaterHoleGoal`.

Introduces a splash effect when water interacts with tiles. The water splash pushes basic nuts away and handles goal interactions.

Adds `INutInteractive` interface to allow tiles to interact with nuts, for use in extensions.