Generalizes water hole goal instantiation
Allows different types of goals to be instantiated as water holes.

The previous implementation was tightly coupled to the water hole goal
type, which limits extensibility. This change allows the type of goal to
be specified in the payload, making the code more flexible and reusable
for other goal types.