Improves character movement event handling.
Adds a 'MoveStarted' event to the character movement sequence, triggered right before smooth movement execution.
This provides a more precise indication of when the character intends to move and allows for actions to be initiated at the start of a move, like resetting a pending data request flag and playing move sound.

Also adds a GridPosition property to Character.