Refactors character movement and nut pushing
Improves character movement logic by separating general movement and nut pushing logic.

Now character first checks if the tile is walkable. If walkable, character moves. If the tile is not walkable, it checks if the tile is a nut. If it is a nut, it checks if the nut can be pushed and pushes it.

Also, allow nut to be pushed to Goal tile.
Removes redundant checks and simplifies the logic for pushing nuts.
Fixes issues where character was not able to push nuts correctly.