Adds user level manager and download button
Implements a user level manager to handle saving user-created levels to the device's local storage.
Adds a download button in the level selection screen, allowing users to manually download the level manual.
The manual is saved to the Downloads folder on Android and a "GetNuts" folder within Documents on Windows.